+++ Tests/SwiftLintFramework/CustomRulesTests.swift
@@ -45,18 +45,66 @@ class CustomRulesTests: XCTestCase {
     }
 
     func testCustomRules() {
-        var regexConfig = RegexConfiguration(identifier: "custom")
-        regexConfig.regex = NSRegularExpression.forcePattern("pattern")
-        regexConfig.matchKinds = Set([SyntaxKind.Comment])
-        var customRuleConfiguration = CustomRulesConfiguration()
-        customRuleConfiguration.customRuleConfigurations = [regexConfig]
-        var customRules = CustomRules()
-        customRules.configuration = customRuleConfiguration
+        let (regexConfig, customRules) = getCustomRules()
+
         let file = File(contents: "// My file with\n// a pattern")
         XCTAssertEqual(customRules.validateFile(file),
                        [StyleViolation(ruleDescription: regexConfig.description,
                         severity: .Warning,
                         location: Location(file: nil, line: 2, character: 6),
                         reason: regexConfig.message)])
+    }
+
+    func testCustomRulesIncludedDefault() {
+        // Violation detected when included is omitted.
+        let (_, customRules) = getCustomRules()
+        let violations = customRules.validateFile(getTestTextFile())
+        XCTAssertEqual(violations.count, 1)
+    }
+
+    func testCustomRulesIncludedExcludesFile() {
+        var (regexConfig, customRules) = getCustomRules(["included": "\\.yml$"])
+
+        var customRuleConfiguration = CustomRulesConfiguration()
+        customRuleConfiguration.customRuleConfigurations = [regexConfig]
+        customRules.configuration = customRuleConfiguration
+
+        let violations = customRules.validateFile(getTestTextFile())
+        XCTAssertEqual(violations.count, 0)
+    }
+
+    func getCustomRules(extraConfig: [String:String] = [:]) -> (RegexConfiguration, CustomRules) {
+        var config = ["regex": "pattern",
+                      "match_kinds": "comment"]
+        extraConfig.forEach { config[$0] = $1 }
+
+        var regexConfig = RegexConfiguration(identifier: "custom")
+        do {
+            try regexConfig.applyConfiguration(config)
+        } catch {
+            XCTFail("Failed regex config")
+        }
+
+        var customRuleConfiguration = CustomRulesConfiguration()
+        customRuleConfiguration.customRuleConfigurations = [regexConfig]
+
+        var customRules = CustomRules()
+        customRules.configuration = customRuleConfiguration
+        return (regexConfig, customRules)
+    }
+
+    func getTestTextFile() -> File {
+        #if SWIFT_PACKAGE
+            let path: String = "Tests/SwiftLintFramework/Resources/test.txt"
+                .absolutePathRepresentation()
+            return File(path: path)!
+        #else
+            let testBundle = NSBundle(forClass: self.dynamicType)
+            let path: String? = testBundle.pathForResource("test", ofType: "txt")
+            if path != nil {
+                return File(path: path!)!
+            }
+        #endif
+        fatalError("Could not load test.txt")
     }
 }